+++ source/Application/Controller/Admin/ShopMain.php
@@ -156,7 +156,6 @@ class ShopMain extends \OxidEsales\Eshop\Application\Controller\Admin\AdminDetai
             'aModules',
             'aModuleControllers',
             'moduleSmartyPluginDirectories',
-            'activeModules',
         ];
         //adding non copable multishop field options
         $multiShopTables = \OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('aMultiShopTables');