+++ tests/Integration/Internal/Framework/Module/Facade/ActiveModulesDataProviderTest.php
@@ -171,11 +171,9 @@ final class ActiveModulesDataProviderTest extends TestCase
     {
         return new ActiveModulesDataProvider(
             $this->get(ShopConfigurationDaoInterface::class),
-            $this->get(ModuleStateServiceInterface::class),
             $this->get(ModulePathResolverInterface::class),
             $this->get(ContextInterface::class),
-            $cache,
-            (new TemplatesDataMapper())
+            $cache
         );
     }